# Conflicts:
#	src/main/java/com/project/linkedindatabase/bootstrap/Bootstrap.java
#	src/main/java/com/project/linkedindatabase/service/modelMap/post/CommentServiceMap.java

+++ src/main/java/com/project/linkedindatabase/controller/ChatController.java
@@ -8,10 +8,13 @@ import com.project.linkedindatabase.service.model.ProfileService;
 import com.project.linkedindatabase.service.model.chat.ChatService;
 import com.project.linkedindatabase.service.model.chat.MessageService;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.Map;
 
 
@@ -41,9 +44,62 @@ public class ChatController {
             chat.setProfileId2(id2);
             chat.setIsArchive(false);
             chat.setMarkUnread(false);
+            if (!chatService.isThereChat(id1,id2))
+                chatService.save(chat);
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+    }
+
+    @CrossOrigin(origins = "*")
+    @PostMapping("/chat-token/{id}")
+    public void createChatById(@RequestHeader Map<String, Object> jsonHeader, @PathVariable long id){
+        String token = JwtUserDetailsService.getTokenByHeader(jsonHeader);
+        Profile profile;
+        try {
+            profile = new JwtUserDetailsService(profileService).getProfileByHeader(jsonHeader);
+
+        }catch (Exception e){
+            e.printStackTrace();
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "There is a problem with token ",e);
+        }
+
+        if (profile.getId() == id)
+        {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "both id is the same ",new Exception("same"));
+        }
+
+        Boolean isThereChat =false;
+        try {
+
+            if (chatService.isThereChat(id,profile.getId()))
+                isThereChat =true;
+        }catch (Exception e){
+            e.printStackTrace();
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "There is a problem with data ",e);
+        }
+
+        if (isThereChat)
+        {
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "There is chat already ",new Exception("duplicate"));
+        }
+
+        try {
+
+            Chat chat = new Chat();
+            chat.setProfileId1(id);
+            chat.setProfileId2(profile.getId());
+            chat.setIsArchive(false);
+            chat.setMarkUnread(false);
             chatService.save(chat);
         }catch (Exception e){
             e.printStackTrace();
+            throw new ResponseStatusException(
+                    HttpStatus.BAD_REQUEST, "There is a problem with token ",e);
         }
     }
 
@@ -54,7 +110,23 @@ public class ChatController {
         try {
             Profile profile = new JwtUserDetailsService(profileService).getProfileByHeader(jsonHeader);
             Chat chat = chatService.findByProfileIds(id1, id2);
-            chatService.deleteByObject(chat);
+            if (chat != null)
+                chatService.deleteByObject(chat);
+
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+    }
+
+    @CrossOrigin(origins = "*")
+    @DeleteMapping("/chat-token/{id}")
+    public void deleteChatById(@RequestHeader Map<String, Object> jsonHeader, @PathVariable long id) throws SQLException {
+        String token = JwtUserDetailsService.getTokenByHeader(jsonHeader);
+        try {
+            Profile profile = new JwtUserDetailsService(profileService).getProfileByHeader(jsonHeader);
+            Chat chat = chatService.findByProfileIds(id, profile.getId());
+            if (chat != null)
+                chatService.deleteByObject(chat);
         }catch (Exception e){
             e.printStackTrace();
         }
@@ -114,7 +186,7 @@ public class ChatController {
 
     @CrossOrigin(origins = "*")
     @GetMapping("/chats/search-user/{searchKey}")
-    public ArrayList<Chat> search(@RequestHeader Map<String, Object> jsonHeader, @PathVariable String searchKey){
+    public ArrayList<Chat> searchForUser(@RequestHeader Map<String, Object> jsonHeader, @PathVariable String searchKey){
         String token = JwtUserDetailsService.getTokenByHeader(jsonHeader);
         try {
             Profile profile = new JwtUserDetailsService(profileService).getProfileByHeader(jsonHeader);
@@ -126,7 +198,7 @@ public class ChatController {
     }
     /*@CrossOrigin(origins = "*")
     @GetMapping("/chats/search-messages/{searchKey}")
-    public ArrayList<Chat> search(@RequestHeader Map<String, Object> jsonHeader, @PathVariable String searchKey){
+    public HashMap<Chat, Message> searchForMessages(@RequestHeader Map<String, Object> jsonHeader, @PathVariable String searchKey){
         String token = JwtUserDetailsService.getTokenByHeader(jsonHeader);
         try {
             Profile profile = new JwtUserDetailsService(profileService).getProfileByHeader(jsonHeader);
